+++ tests/contracts/BlocklistVerifier.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.0;
+
+import {BlocklistVerifier as Verifier} from "@zrclib/sdk/contracts/generated/BlocklistVerifier.sol";
+
+contract BlocklistVerifier is Verifier {}
