--- tests/contracts/ComplianceVerifier.sol
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-pragma solidity ^0.8.0;
-
-import {BlocklistVerifier} from "@zrclib/sdk/contracts/generated/BlocklistVerifier.sol";
-
-contract ComplianceVerifier is BlocklistVerifier {}
