+++ tests/contracts/CompliantShieldedPool.sol
@@ -1,11 +1,14 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.9;
 
-import {MerkleTreeForBlocklist} from "@zrclib/sdk/contracts/MerkleTreeForBlocklist.sol";
+import {ShieldedPoolWithBlocklist} from "@zrclib/sdk/contracts/ShieldedPoolWithBlocklist.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
-contract CompliantShieldedPool is MerkleTreeForBlocklist {
+contract CompliantShieldedPool is ShieldedPoolWithBlocklist {
     constructor(
-        address _verifier
-    ) MerkleTreeForBlocklist(5, _verifier) {}
+        address _hasher,
+        address _transactionVerifier,
+        address _blocklistVerifier,
+        address _swapExecutor
+    ) ShieldedPoolWithBlocklist(5, _hasher, _transactionVerifier, _blocklistVerifier, _swapExecutor) {}
 }
